@@ -10,8 +10,7 @@
         }
     ],
     "bin": [
-        "bin/asynit",
-        "bin/smoker"
+        "bin/asynit"
     ],
     "autoload": {
         "psr-4": {
@@ -27,24 +26,31 @@
         "test": "bin/asynit tests/"
     },
     "require": {
-        "php": "^7.3 || ^8.0",
-        "amphp/http-client": "^3.0.14",
-        "amphp/sync": "^1.0",
-        "bovigo/assert": "^6.2",
-        "doctrine/annotations": "^1.2",
-        "doctrine/collections": "^1.3",
-        "guzzlehttp/psr7": "^1.4",
-        "php-http/client-common": "^2.0",
-        "php-http/message": "^1.3",
+        "php": "^8.1",
+        "amphp/amp": "^3.0",
+        "amphp/sync": "^2.0",
+        "bovigo/assert": "^7.0",
         "symfony/console": "^4.4 || ^5.0 || ^6.0",
-        "symfony/dom-crawler": "^4.4 || ^5.0 || ^6.0",
-        "symfony/finder": "^4.4 || ^5.0 || ^6.0",
-        "symfony/yaml": "^4.4 || ^5.0 || ^6.0"
+        "symfony/finder": "^4.4 || ^5.0 || ^6.0"
+    },
+    "suggest": {
+        "amphp/http-client": "To use the web test case trait",
+        "nyholm/psr7": "To use the web test case trait"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^2.10"
+        "amphp/http-client": "v5.0.0-beta.11",
+        "friendsofphp/php-cs-fixer": "^v2.19.3",
+        "nyholm/psr7": "^1.8.0",
+        "php-http/client-common": "^2.4",
+        "php-http/message": "^1.3",
+        "phpstan/phpstan": "^1.10"
+    },
+    "conflict": {
+        "amphp/http-client": "<v5.0.0",
+        "symfony/options-resolver": "<4.4.30 || <5.3 >= 5.0"
     },
     "prefer-stable": true,
+    "minimum-stability": "dev",
     "config": {
         "sort-packages": true
     }
