@@ -2,14 +2,11 @@
 
 namespace Asynit\Command;
 
-use Amp\Loop;
 use Asynit\Output\OutputFactory;
 use Asynit\Parser\TestPoolBuilder;
 use Asynit\Parser\TestsFinder;
 use Asynit\Runner\PoolRunner;
 use Asynit\TestWorkflow;
-use Doctrine\Common\Annotations\AnnotationReader;
-use Http\Message\MessageFactory\GuzzleMessageFactory;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -55,16 +52,12 @@ class AsynitCommand extends Command
 
         list($chainOutput, $countOutput) = (new OutputFactory($input->getOption('order')))->buildOutput(\count($testMethods));
 
-        // Build services for parsing and running tests
-        $builder = new TestPoolBuilder(new AnnotationReader());
-        $runner = new PoolRunner(new GuzzleMessageFactory(), new TestWorkflow($chainOutput), $input->getOption('concurrency'));
+        $builder = new TestPoolBuilder();
+        $runner = new PoolRunner(new TestWorkflow($chainOutput), $input->getOption('concurrency'));
 
         // Build a list of tests from the directory
         $pool = $builder->build($testMethods);
-
-        Loop::run(function () use ($runner, $pool) {
-            $runner->loop($pool);
-        });
+        $runner->loop($pool);
 
         // Return the number of failed tests
         return $countOutput->getFailed();
