@@ -2,19 +2,48 @@
 
 namespace Asynit\Tests;
 
-use function Amp\delay;
 use Asynit\Attribute\Depend;
 use Asynit\Attribute\TestCase;
-use Asynit\Assert\AssertCaseTrait;
+use Asynit\HttpClient\HttpClientWebCaseTrait;
+use Psr\Http\Message\ResponseInterface;
 
 #[TestCase]
-class FunctionalTests
+class FunctionalHttpTests
 {
-    use AssertCaseTrait;
+    use HttpClientWebCaseTrait;
 
     public function testReturn()
     {
         return 'tata';
+    }
+
+    public function testGet()
+    {
+        $response = $this->get($this->createUri('/'));
+
+        $this->assertInstanceOf(ResponseInterface::class, $response);
+        $this->assertStatusCode(200, $response);
+
+        return 'foo';
+    }
+
+    public function testException()
+    {
+        $exception = null;
+
+        try {
+            $response = $this->get('http://something-is-not-reachable');
+        } catch (\Exception $e) {
+            $exception = $e;
+        }
+
+        $this->assertNotNull($exception, 'Not null exception');
+    }
+
+    #[Depend('testGet')]
+    public function testDepend($value)
+    {
+        $this->assertSame('foo', $value);
     }
 
     #[Depend("Asynit\Tests\AnotherTest::test_from_another_file")]
@@ -30,22 +59,22 @@ class FunctionalTests
 
     public function testParallel1()
     {
-        delay(4);
+        $this->get($this->createUri('/delay/1'));
     }
 
     public function testParallel2()
     {
-        delay(5);
+        $this->get($this->createUri('/delay/3'));
     }
 
     public function testParallel3()
     {
-        delay(6);
+        $this->get($this->createUri('/delay/5'));
     }
 
     public function testParallel4()
     {
-        delay(7);
+        $this->get($this->createUri('/delay/7'));
     }
 
     #[Depend('testStartParallel')]
@@ -89,5 +118,10 @@ class FunctionalTests
         $this->assertSame('a', $a);
         $this->assertSame('b', $b);
         $this->assertSame('d', $d);
+    }
+
+    protected function createUri(string $uri): string
+    {
+        return 'http://127.0.0.1:8081'.$uri;
     }
 }
